requirements.txt
# Django and Django REST Framework
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-environ>=0.11.0

# Database
psycopg2-binary>=2.9.0

# Celery and Redis
celery>=5.3.0
redis>=5.0.0

# Scientific computing (for DBSCAN clustering)
scikit-learn>=1.3.0
shapely>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Production WSGI server
gunicorn>=21.2.0


